requirements.txt
aiogram
bcrypt
python-dotenv
flask
flask-sqlalchemy
flask-migrate
flask-marshmallow
marshmallow-sqlalchemy
flask-cors
email-validator
werkzeug
Flask-Mail
requests
flasgger
flask-socketio
eventlet
cryptography
stripe
pytz
